% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{bar plot function}
\usage{
bar(
  file,
  x_column,
  y_column,
  horizontal = TRUE,
  Stacked = FALSE,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{x_column}{Name of the column from the file you want to plot on the x-axis}

\item{y_column}{Name of the column from the file you want to plot on the y-axis}

\item{horizontal}{Put True if you want to plot a horizontal bar plot, False for vertical plot.}

\item{primary_color}{First color for bar plot}

\item{secondary_color}{Second color for bar plot}

\item{tertiary_color}{Third color for bar plot}

\item{quaternary_color}{Fourth color for bar plot}

\item{quinary_color}{Fifth color for bar plot}

\item{senary_color}{Sixth color for bar plot}

\item{stacked}{Put True for a stacked plot, False for normal bar plot}
}
\value{
A bar plot
}
\description{
Creates a bar plot of the data.
}
\details{
This function creates a bar plot based on two columns of a file with data.
}
\examples{
  \dontrun{
     bar("testdata.csv", 
"column_1" 
"column2", 
TRUE, FALSE, 
"#004D9A", 
"#002142", 
"#0069D2", 
"#0180FF",
"#004D9A", 
"#002142") )
  }

}
\author{
Emiel Creus
}
