% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Box plot function}
\usage{
boxplot(
  file,
  columns,
  mean_dot_symbol,
  mean_dot = FALSE,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{columns}{Names of the column from the file you want to make a boxplot of.}

\item{mean_dot_symbol}{Number representing the shape of the symbol for the mean dot.}

\item{mean_dot}{Put True if you want to plot the mean dot.}

\item{primary_color}{First color for bar plot}

\item{secondary_color}{Second color for bar plot}

\item{tertiary_color}{Third color for bar plot}

\item{quaternary_color}{Fourth color for bar plot}

\item{quinary_color}{Fifth color for bar plot}

\item{senary_color}{Sixth color for bar plot}
}
\value{
A box plot
}
\description{
Creates a box plot of the data.
}
\details{
This function creates a box plot based on a column of a file with data.
}
\examples{
  \dontrun{
     boxplt("testdata.csv",
c("column_1","column_2"),
19, TRUE,
"#004D9A",
"#002142",
"#0069D2", "#0180FF",
"#004D9A", "#002142") )
  }

}
\author{
Emiel Creus
}
