% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp}
\alias{bp}
\title{generic businessplot function}
\usage{
bp(file)
}
\arguments{
\item{file}{A csv file with the source data}
}
\value{
A graphical plot
}
\description{
Creates a general plot of the data.
}
\details{
This function creates a plot based on the format of the data
and this can be: scatterplot, lineplot or bar plot.
}
\examples{
  \dontrun{
     bp("testdata.csv")
  }


}
\author{
Yves R. Sagaert

at import forecast
at import Mcomp
}
