% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessplots_package.R
\docType{package}
\name{businessplots}
\alias{businessplots}
\alias{businessplots-package}
\title{businessplots: A package for performing basic plots for business}
\description{
This package provides functionality for basic business plots
with a standardised way of color scheme.
}
\author{
Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
