% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{create_treemap}
\alias{create_treemap}
\title{generic businessplot function}
\usage{
create_treemap(
  file,
  index,
  quantity,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{index}{Name of the index column from the file you want to plot}

\item{quantity}{Put True if you want to plot a horizontal bar plot, False for vertical plot.}

\item{primary_color}{First color for bar plot}

\item{secondary_color}{Second color for bar plot}

\item{tertiary_color}{Third color for bar plot}

\item{quaternary_color}{Fourth color for bar plot}

\item{quinary_color}{Fifth color for bar plot}

\item{senary_color}{Sixth color for bar plot}

\item{y_column}{Name of the quantity column from the file you want to plot}
}
\value{
A graphical plot
}
\description{
Creates a general plot of the data.
}
\details{
This function creates a plot based on the format of the data
and this can be: treemap
}
\examples{
  \dontrun{
     create_treemap("testdata.csv", "Year", "Quota", "#004D9A", "#002142", "#0069D2", "#0180FF", "#004D9A", "#002142")
  }



}
\author{
Isolde Boussauw

at import forecast
at import Mcomp
}
