% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualaxesscatter.R
\name{dualaxscatter}
\alias{dualaxscatter}
\title{multiple line scatter plot function with multiple scales}
\usage{
dualaxscatter(
  file,
  x_column,
  y_columns,
  legend_pos,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{x_column}{Name of the column from the file you want to plot on the x-axis}

\item{y_columns}{Name of the columns from the file you want to plot on the y-axes}

\item{legend_pos}{the position of the legend}

\item{primary_color}{First color for scatter plot}

\item{secondary_color}{Second color for scatter plot}

\item{tertiary_color}{Third color for scatter plot}

\item{quaternary_color}{Fourth color for scatter plot}

\item{quinary_color}{Fifth color for scatter plot}

\item{senary_color}{Sixth color for scatter plot}
}
\value{
A multiple line plot with multiple scales
}
\description{
Creates a plot of the data with multiple scatter plots that have a different scale.
}
\details{
This function creates a plot with multiple scatter plots and a different scale based on the format of the data
}
\examples{
  \dontrun{
     dualaxscatter("Testdata.csv", "Year", c("Quota", "Incidents","Complatins"),"topleft","red","green","orange","yellow","blue","black")
  }


}
\author{
Emiel Creus
}
