% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexdashboardmaker.R
\name{flexdashboardmaker}
\alias{flexdashboardmaker}
\title{Flexdashboard maker}
\usage{
flexdashboardmaker()
}
\description{
Creates a flexdashboard automatically.
}
\details{
This function creates a flexdashboard automatically.
}
\examples{
  \dontrun{
     flexdashboardmaker()
  }



}
\author{
Ruben Vanhecke
}
