% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastplot.R
\name{forecastplot}
\alias{forecastplot}
\title{gauge chart function}
\usage{
forecastplot(file = "ProductionDataSample.csv", X, y)
}
\arguments{
\item{file}{A csv file with source data}

\item{X}{The column name of the actual values}

\item{y}{The column name of the predicted values}
}
\value{
A forecast plot
}
\description{
Creates a forecast plot based the actual values and the predicted values.
}
\details{
This function creates a forecast plot based on the actual values and the predicted values.
}
\examples{
  \dontrun{
       forecastplot("ProductionDataSample.csv", 
"Weight", 
"WeightPredicted")
  }



}
\author{
Ruben Vanhecke

at import greybox
}
