% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatobject.R
\name{formatobject}
\alias{formatobject}
\title{formatobject function}
\usage{
formatobject(file = "formatobject.csv")
}
\arguments{
\item{file}{A csv file with the source data}
}
\value{
Styles
}
\description{
Retrieve the styles for all the charts.
}
\details{
This function retrieves all the styles (like colors, fonts, line thicknness...) for all the charts.
}
\examples{
\dontrun{
 formatobject("testdata.csv")
 extract_data_formatobject(formatobject("testdata.csv"), "colors")
}



}
\author{
Ruben Vanhecke
}
