% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funmaker.R
\name{funmaker}
\alias{funmaker}
\title{Function maker}
\usage{
funmaker()
}
\description{
Creates a function automatically (without body).
}
\details{
This function creates a function automatically (without body).
}
\examples{
  \dontrun{
     funmaker()
  }



}
\author{
Yves R.Sagaert, Ruben Vanhecke
}
