% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugechart.R
\name{gaugechart}
\alias{gaugechart}
\title{gauge chart function}
\usage{
gaugechart(
  gauge_value,
  max_value,
  type,
  mode,
  title,
  title_size,
  axis_tick_width,
  border_width,
  threshold_width,
  threshold_thickness,
  font_family,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{gauge_value}{An integer or a numeric value}

\item{max_value}{An integer or a numeric value}

\item{type}{The type of the gauge chart}

\item{mode}{The mode of the gauge chart}

\item{title}{The title of the gauge chart}

\item{title_size}{The size of the title of the gauge chart}

\item{axis_tick_width}{The tick width of the axis of the gauge chart}

\item{border_width}{The border width of the gauge chart}

\item{threshold_width}{The threshold width}

\item{threshold_thickness}{The thickness of the threshold}

\item{font_family}{The font family of the text in the gauge chart}

\item{primary_color}{The primary color}

\item{secondary_color}{The secondary color}

\item{tertiary_color}{The tertiary color}

\item{quaternary_color}{The quaternary color}

\item{quinary_color}{The quinary color}

\item{senary_color}{The senary color}
}
\value{
A gauge chart
}
\description{
Creates a gauge chart of an integer or a numeric value.
}
\details{
This function creates a gauge chart based on an integer or a numeric value.
}
\examples{
  \dontrun{
     gaugechart(25, 100, "indicator", "gauge+number", "Visitors inside", 24, 1, 2, 4, 0.75, "Arial",
       "#004D9A", "#002142", "#0069D2", "#0180FF", "#4FA7FF", "#A7D3FF")
  }



}
\author{
Ruben Vanhecke

at import ggplot2
at import plotly
}
