% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histogram}
\alias{histogram}
\title{histogram plot function}
\usage{
histogram(
  file,
  main_column,
  show_normal_distribution = FALSE,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{main_column}{Name of the column from the file you want to plot in the histogram}

\item{show_normal_distribution}{Put True if you want to show the normal distribution on top of the histogram}

\item{primary_color}{First color for bar plot}

\item{secondary_color}{Second color for bar plot}

\item{tertiary_color}{Third color for bar plot}

\item{quaternary_color}{Fourth color for bar plot}

\item{quinary_color}{Fifth color for bar plot}

\item{senary_color}{Sixth color for bar plo}
}
\value{
A histogram plot
}
\description{
Creates a histogram of the data.
}
\details{
This function creates a histogram plot based on a column of a file with data.
}
\examples{
  \dontrun{
     histogram("testdata.csv", "column_1", TRUE, "#004D9A", "#002142", "#0069D2", "#0180FF","#004D9A", "#002142")
  }

}
\author{
Emiel Creus
}
