% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownmaker.R
\name{markdownmaker}
\alias{markdownmaker}
\title{Markdown maker}
\usage{
markdownmaker()
}
\description{
Creates a markdown automatically.
}
\details{
This function creates a markdown automatically.
}
\examples{
  \dontrun{
     markdownmaker()
  }



}
\author{
Ruben Vanhecke
}
