% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplescaleline.R
\name{multiscaleline}
\alias{multiscaleline}
\title{multipleline plot function with multiple scales}
\usage{
multiscaleline(
  file,
  x_column,
  y_columns,
  legend_pos,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{x_column}{Name of the column from the file you want to plot on the x-axis}

\item{y_columns}{Name of the columns from the file you want to plot on the y-axes}

\item{legend_pos}{the position of the legend}

\item{primary_color}{First color for line plot}

\item{secondary_color}{Second color for line plot}

\item{tertiary_color}{Third color for line plot}

\item{quaternary_color}{Fourth color for line plot}

\item{quinary_color}{Fifth color for line plot}

\item{senary_color}{Sixth color for line plot}
}
\value{
A multiple line plot with multiple scales
}
\description{
Creates a plot of the data with multiple lines that have a different scale.
}
\details{
This function creates a plot with multiple lines and a different scale based on the format of the data
}
\examples{
  \dontrun{
     multiscaleline("testdata.csv", column_1 , c("column2", "column3", "column4"), "topleft", "red","green","orange","yellow","blue","black")
  }


}
\author{
Emiel Creus
}
