% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscatter.R
\name{multiscatter}
\alias{multiscatter}
\title{multiple scatter plot function}
\usage{
multiscatter(
  file,
  x_column,
  y_column,
  legend_pos,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{x_column}{Name of the column from the file you want to plot on the x-axis}

\item{y_column}{Name of the columns from the file you want to plot on the y-axes}

\item{legend_pos}{the position of the legend}

\item{primary_color}{First color for scatter plot}

\item{secondary_color}{Second color for scatter plot}

\item{tertiary_color}{Third color for scatter plot}

\item{quaternary_color}{Fourth color for scatter plot}

\item{quinary_color}{Fifth color for scatter plot}

\item{senary_color}{Sixth color for scatter plot}
}
\value{
A multiple scatter plot
}
\description{
Creates a plot of the data with multiple scatter plots
}
\details{
This function creates a plot with multiple scatter plots based on the format of the data
}
\examples{
  \dontrun{
     multiscatter("testdata.csv",
column_1 ,
c("column2", "column3", "column4"),
"topleft", "red","green","
orange","yellow","blue","black")
  }


}
\author{
Emiel Creus
}
