% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart.R
\name{piechart}
\alias{piechart}
\title{pie chart function}
\usage{
piechart(
  file,
  main,
  pie_label,
  pie_data_label,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{main}{Title of the pie chart}

\item{pie_label}{Name of the column of the file with the labels of the pie chart}

\item{pie_data_label}{Name of the column of the file with the data of the pie chart}

\item{primary_color}{First color for pie chart}

\item{secondary_color}{Second color for pie chart}

\item{tertiary_color}{Third color for pie chart}

\item{quaternary_color}{Fourth color for pie chart}

\item{quinary_color}{Fifth color for pie chart}

\item{senary_color}{Sixth color for pie chart}
}
\value{
A pie chart
}
\description{
Creates a pie chart of the data.
}
\details{
This function creates a pie chart based on a column of a file with data.
}
\examples{
  \dontrun{
     piechart(
"testdata.csv", 
"Year", "Quota", 
"#004D9A", "#002142", 
"#0069D2", "#0180FF", 
"#4FA7FF", "#A7D3FF")
  }



}
\author{
Ruben Vanhecke
}
