% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_letters_with_numbers.R
\name{replace_letters_with_numbers}
\alias{replace_letters_with_numbers}
\title{Replace letters with numbers}
\usage{
replace_letters_with_numbers(df, column_name)
}
\arguments{
\item{df}{Dataframe to replace letters of one column to numbers.}

\item{column_name}{Specify the name of the column in which the replacement
needs to happen.}
}
\value{
A dataframe
}
\description{
Function to replace letters with numbers
}
\details{
This function replace letters with numbers for one column of a dataframe.
}
\examples{
  \dontrun{
    data <- data.frame(
     id_column = c("ID1","ID2","ID3","ID4"),
     score = c("A","B","C","D"))
    replace_letters_with_numbers(data,column_name=score)
  }

}
\author{
Yves R. Sagaert
}
