% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatter plot function}
\usage{
scatter(
  file,
  x_column,
  y_column,
  symbol,
  auto_fit = TRUE,
  primary_color,
  secondary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{x_column}{Name of the column from the file you want to plot on the x-axis}

\item{y_column}{Name of the column from the file you want to plot on the y-axis}

\item{symbol}{number representing the symbol used for the scatter plot}

\item{auto_fit}{Put false to start from 0, True shows the best fit.}

\item{primary_color}{First color for bar plot}

\item{secondary_color}{Second color for bar plot}
}
\value{
A scatter plot
}
\description{
Creates a scatter plot of the data.
}
\details{
This function creates a plot based on the format of the data
}
\examples{
  \dontrun{
      scatter("testdata.csv", 
"column_1", "column2", 
2, FALSE, 
"#004D9A", "#002142")
  }


}
\author{
Emiel Creus
}
