% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard.R
\name{shinydashboard}
\alias{shinydashboard}
\title{Shiny Dashboard function}
\usage{
shinydashboard()
}
\description{
Creates a Shiny Dashboard.
}
\details{
This function creates a Shiny Dashboard.
}
\examples{
  \dontrun{
     shinydashboard()
  }

}
\author{
Ruben Vanhecke

at import shiny
at import shinydashboard
}
