% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderchart.R
\name{spiderchart}
\alias{spiderchart}
\title{spider chart function}
\usage{
spiderchart(
  file,
  spider_label,
  spider_data_label,
  title,
  polygon_line_width,
  net_line_type,
  net_line_width,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{spider_label}{Name of the column of the file with the labels of the spider chart}

\item{spider_data_label}{Name of the column of the file with the data of the spider chart}

\item{title}{Title of the spider chart}

\item{polygon_line_width}{Line width of the polygon}

\item{net_line_type}{Net line type}

\item{net_line_width}{Net width}

\item{primary_color}{First color for spider chart}

\item{secondary_color}{Second color for spider chart}

\item{tertiary_color}{Third color for spider chart}

\item{quaternary_color}{Fourth color for spider chart}
}
\value{
A spider chart
}
\description{
Creates a spider chart of the data.
}
\details{
This function creates a spider chart based on two columns of a file with data.
}
\examples{
  \dontrun{
     spiderchart("testdata.csv", "Year", "Quota", "Annual quota", 4, 2, 2, "#004D9A", "#002142", "#0069D2", "#0180FF")
  }



}
\author{
Ruben Vanhecke

at import fmsb
}
