% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_plot.R
\name{stability_plot}
\alias{stability_plot}
\title{stability plot function}
\usage{
stability_plot(a, b, c, percentage)
}
\arguments{
\item{a}{Metric on top of the central circle. For the stability plot this
represents the overall upwards lift.}

\item{b}{Metric below the central circle. For the stability plot this
represents the overall downwards lift.}

\item{c}{Metric on the horizontal axis (+/-). For the stability plot this
represents the overall bias.}

\item{percentage}{Percent stability, noted as mark in the circle.}
}
\value{
A visualisation plot
}
\description{
Creates a custom designed visualisation of different metrics.
}
\details{
This function creates a custom designed visualisation of the metric on
stabilisation of the model.
}
\examples{
  \dontrun{
     stability_plot(a=3, b=6, c=10, percentage=25)
  }


}
\author{
Yves R. Sagaert
}
