% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_map_plot.R
\name{world_map_plot}
\alias{world_map_plot}
\title{world map plot}
\usage{
world_map_plot(data, modelname = "Model X", white = TRUE)
}
\arguments{
\item{data}{A dataframe with at least 2 columns: ISO code of the country and
a color for the country.}

\item{modelname}{Label for the plot of the world map}

\item{white}{Boolean by default set on 'TRUE' to indicate the color of
countries with value NA, or not included in data for the score.}
}
\value{
A plot
}
\description{
Creates a plot of world with color coding.
}
\details{
This function creates a plot of a map with different colors for different countries.
}
\examples{
  \dontrun{
    data <- data.frame(
     country_code = c("USA", "CAN", "BRA", "FRA", "DEU", "IND", "CHN"),
     score = c(1, 2, 3, 4, 5, 6, 7))
    world_map_plot(data)
  }

}
\author{
Yves R. Sagaert
}
