% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barwithline.R
\name{barwithline}
\alias{barwithline}
\title{function for bar plot with line}
\usage{
barwithline(
  file,
  x_column,
  y_columns,
  line_type,
  legend_pos,
  primary_color,
  secondary_color,
  tertiary_color,
  quaternary_color,
  quinary_color,
  senary_color
)
}
\arguments{
\item{file}{A csv file with the source data}

\item{x_column}{Name of the column from the file you want to plot on the x-axis}

\item{line_type}{The type of line you want for the line plot}

\item{primary_color}{First color for line plot}

\item{secondary_color}{Second color for line plot}

\item{y_column}{Name of the column from the file you want to plot on the y-axis}

\item{symbol}{Number of the symbol you want for the plot}

\item{auto_fit}{Put false to start from 0, True shows the best fit.}
}
\value{
A bar plot with line
}
\description{
Creates a bar plot with line
}
\details{
this fucntion plots a bar plot with line based on the data
}
\examples{
  \dontrun{
     barwithline("testdata.csv", 
column_1, 
c("column2", "column3", "column4"), 
1, 1, "red","green",
"orange","yellow","
blue","black")
  }


}
\author{
Isolde Boussauw
}
